back-end/requirements.txt
Django==5.1.4
PyJWT==2.10.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
uvicorn==0.32.1
drf-yasg==1.21.8